ai_models/flower-ai/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
opencv-python-headless
tensorflow
h5py
